+++ packages/field/src/components/Select/index.tsx
@@ -8,6 +8,7 @@ import React, {
   useCallback,
   useEffect,
 } from 'react';
+import type { SelectProps } from 'antd';
 import { Space, Spin, ConfigProvider } from 'antd';
 import type {
   ProFieldRequestData,
@@ -17,7 +18,12 @@ import type {
   RequestOptionsType,
 } from '@ant-design/pro-utils';
 
-import { useDebounceFn, useDeepCompareEffect, useMountMergeState } from '@ant-design/pro-utils';
+import {
+  useDebounceFn,
+  nanoid,
+  useDeepCompareEffect,
+  useMountMergeState,
+} from '@ant-design/pro-utils';
 import useSWR from 'swr';
 import { useIntl } from '@ant-design/pro-provider';
 
@@ -27,8 +33,6 @@ import type { ProFieldStatusType } from '../Status';
 import TableStatus, { ProFieldBadgeColor } from '../Status';
 import type { ProFieldFC } from '../../index';
 import './index.less';
-
-let testId = 0;
 
 type SelectOptionType = Partial<RequestOptionsType>[];
 
@@ -262,8 +266,7 @@ export const useFieldFetchData = (
       return props.proFieldKey.toString();
     }
     if (props.request) {
-      testId += 1;
-      return testId.toString();
+      return nanoid();
     }
     return 'no-fetch';
   });
@@ -309,7 +312,7 @@ export const useFieldFetchData = (
     },
     [],
     props.debounceTime ?? 0,
-    // 因为使用了swc，自动清理请求可能导致缓存错误的数据
+    // 因为使用了swr，自动清理请求可能导致缓存错误的数据
     true,
   );
 
@@ -385,7 +388,10 @@ export const useFieldFetchData = (
  *
  * @param
  */
-const FieldSelect: ProFieldFC<FieldSelectProps> = (props, ref) => {
+const FieldSelect: ProFieldFC<FieldSelectProps & Pick<SelectProps, 'fieldNames'>> = (
+  props,
+  ref,
+) => {
   const {
     mode,
     valueEnum,
@@ -407,10 +413,13 @@ const FieldSelect: ProFieldFC<FieldSelectProps> = (props, ref) => {
   const inputRef = useRef();
   const intl = useIntl();
   const keyWordsRef = useRef<string>('');
+  const { fieldNames } = fieldProps;
 
-  useEffect(() => {
-    testId += 1;
-  }, []);
+  const {
+    label: labelPropsName = 'label',
+    value: valuePropsName = 'value',
+    options: optionsPropsName = 'options',
+  } = fieldNames || {};
 
   useEffect(() => {
     keyWordsRef.current = fieldProps?.searchValue;
@@ -423,24 +432,30 @@ const FieldSelect: ProFieldFC<FieldSelectProps> = (props, ref) => {
     fetchData: () => fetchData(),
   }));
 
+  const optionsValueEnum = useMemo<Record<string, any>>(() => {
+    const traverseOptions = (_options: typeof options): Record<string, any> => {
+      return _options?.length > 0
+        ? _options?.reduce((pre, cur) => {
+            const curLabel = cur[labelPropsName],
+              curValue = cur[valuePropsName],
+              curOptions = cur[optionsPropsName];
+            return {
+              ...pre,
+              [curValue]: curLabel,
+              ...traverseOptions(curOptions),
+            };
+          }, {})
+        : {};
+    };
+    return traverseOptions(options);
+  }, [labelPropsName, options, optionsPropsName, valuePropsName]);
+
   if (mode === 'read') {
-    const optionsValueEnumMap: ProSchemaValueEnumMap = new Map();
-    options?.forEach((opt) => {
-      optionsValueEnumMap.set(opt.value ?? '', opt.label);
-    });
-
-    // 如果有 label 直接就用 label
-    // @ts-ignore
-    if (rest.text?.label) {
-      // @ts-ignore
-      return rest.text?.label;
-    }
-
     const dom = (
       <>
         {proFieldParsingText(
           rest.text,
-          ObjToMap(valueEnum || optionsValueEnumMap) as unknown as ProSchemaValueEnumObj,
+          ObjToMap(valueEnum || optionsValueEnum) as unknown as ProSchemaValueEnumObj,
         )}
       </>
     );