+++ packages/form/src/BaseForm/index.tsx
@@ -159,7 +159,7 @@ function BaseForm<T = Record<string, any>>(props: BaseFormProps<T>) {
     onReset,
     omitNil = true,
     isKeyPressSubmit,
-    autoFocusFirstInput,
+    autoFocusFirstInput = true,
     ...rest
   } = props;
   const [inlineForm] = Form.useForm(form);
@@ -206,6 +206,7 @@ function BaseForm<T = Record<string, any>>(props: BaseFormProps<T>) {
         const values = await formRef.current?.validateFields(nameList);
         return transformKey(values, omitNil);
       },
+      formRef,
     }),
     [omitNil, transformKey],
   );